Move data fetch to DataService
