+++ hosting/src/app/app.component.ts
@@ -7,6 +7,7 @@ import {Auth, user} from '@angular/fire/auth';
 import {map, Observable} from 'rxjs';
 import {WeekTableComponent} from './week-table.component';
 import {BookableUnit, ConfigData, PricingTier, Reservation, ReservableWeek} from './types';
+import {DataService} from './data-service';
 
 
 @Component({
@@ -29,55 +30,15 @@ export class AppComponent {
 
   title = 'Reservations-App';
 
-  // Eventually, this will be dynamic…
-  configYear = 2025;
   weeks$: Observable<ReservableWeek[]>;
   reservations$: Observable<Reservation[]>;
   units$: Observable<BookableUnit[]>;
   pricingTiers$: Observable<{ [key: string]: PricingTier }>;
 
-  constructor() {
-    // Get the bookable unit documents … with the ID field.
-    const bookableUnitsCollection = collection(this.firestore, 'units').withConverter<BookableUnit>({
-      fromFirestore: snapshot => {
-        const {name} = snapshot.data();
-        const {id} = snapshot;
-        return {id, name};
-      },
-      toFirestore: (it: any) => it,
-    });
-    this.units$ = collectionData(bookableUnitsCollection).pipe();
-
-    // Get the pricing tier documents … with the ID field.
-    // Also, store as a map from id to pricing tier.
-    const pricingTiersCollection = collection(this.firestore, 'pricingTiers').withConverter<PricingTier>({
-      fromFirestore: snapshot => {
-        const {name, color} = snapshot.data();
-        const {id} = snapshot;
-        return {id, name, color};
-      },
-      toFirestore: (it: any) => it,
-    });
-    this.pricingTiers$ = collectionData(pricingTiersCollection).pipe(
-      map(
-        it => {
-          return it.reduce((acc, tier) => {
-            acc[tier.id] = tier;
-            return acc;
-          }, {} as { [key: string]: PricingTier });
-        }
-      )
-    );
-
-    const weeksCollection = collection(this.firestore, 'weeks');
-    const weeksQuery = query(weeksCollection, where('year', '==', this.configYear), limit(1));
-    this.weeks$ = collectionData(weeksQuery).pipe(
-      map((it) => it[0] as ConfigData),
-      map((it) => it?.weeks || []),
-    );
-
-    const reservationsCollection = collection(this.firestore, 'reservations');
-    const reservationsQuery = query(reservationsCollection, where('startDate', '>=', String(this.configYear)), where('endDate', '<', String(this.configYear + 1)));
-    this.reservations$ = collectionData(reservationsQuery).pipe() as Observable<Reservation[]>;
+  constructor(dataService: DataService) {
+    this.pricingTiers$ = dataService.pricingTiers$;
+    this.reservations$ = dataService.reservations$;
+    this.units$ = dataService.units$;
+    this.weeks$ = dataService.weeks$;
   }
 }
